backend/requirements.txt
django>=5.0
djangorestframework>=3.14
pandas>=2.1
